Refactor dbt adapter to dbt util
