--- piperider_cli/adapter/__init__.py
@@ -1,1 +1,0 @@
-from .dbt_adapter import DbtAdapter, DbtAdaptee
